Fix up aica firmware licensing
The 3-clause BSD that this firmware uses requires more than just
keeping the copyright notice, like license.txt currently says.
Rather, it has has that "redistributions in binary form (which is
what the ALSA Project is doing, since only the firmware binary is
provided) must reproduce the above copyright notice, this list of
conditions and the following disclaimer in the documentation and/or
other materials provided with the distribution." This means that more
than just the copyright notice needs to be included. This change is
intended to address this by adding the parts that were omitted
originally.

Signed-off-by: Jason Self <j@jxself.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>